Only use specific language tags in EoL Articles from EoL that do not have an explicit language tag are rejected. Language filtering moved to ExternalSiteService rather than page, since it's all going through there anyway.
